updated imports, and nuked utils folder, moved remaining files into service folder
