+++ backend/service/aiService.js
